tests/wpt/web-platform-tests/forced-colors-mode/forced-colors-mode-32.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Forced colors mode - webkit-tap-highlight-color with sys color.</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-colors-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  a {
    -webkit-tap-highlight-color: GrayText;
  }
</style>
<a href="https://www.wikipedia.org" id="link">
  This link color should not be overridden when forced colors mode is enabled.
</a>
<a href="" id="ref" style="forced-color-adjust:none"></a>

<script>
  var gray_text = getComputedStyle(ref).webkitTapHighlightColor;
  test(function(){
    assert_equals(getComputedStyle(link).webkitTapHighlightColor, gray_text);
  }, "Checks hyperlinks are not overridden if a sys color is used.");
</script>
